# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lowo Control Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/embeds/{id}:
    get:
      operationId: getEmbed
      tags: [embeds]
      summary: Get a single embed override
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Embed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEmbed
      tags: [embeds]
      summary: Update an embed override
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbedData"
      responses:
        "200":
          description: Updated embed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedEntry"
    delete:
      operationId: deleteEmbed
      tags: [embeds]
      summary: Reset embed to default (delete override)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Modules ───────────────────────────────────────────────────────────────
...
  /dashboard/files/content:
    get:
      operationId: getFileContent
      tags: [files]
      summary: Get file content (secrets redacted)
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "403":
          description: Forbidden path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/files/search:
    get:
      operationId: searchFiles
      tags: [files]
      summary: Search file content
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileSearchResult"

  # ── Database ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: ["string", "null"]
...
    EmbedEntry:
      type: object
      required: [id, module, title, description]
      properties:
        id:
          type: string
        module:
          type: string
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        color:
          type: ["integer", "null"]
        footer:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        fields:
          type: array
          items:
            $ref: "#/components/schemas/EmbedField"
        isDefault:
          type: boolean
        lastModified:
          type: ["string", "null"]
        lastModifiedBy:
          type: ["string", "null"]
    EmbedField:
      type: object
      required: [name, value, inline]
      properties:
        name:
          type: string
        value:
          type: string
        inline:
          type: boolean
    EmbedData:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        color:
          type: ["integer", "null"]
        footer:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        fields:
          type: array
          items:
            $ref: "#/components/schemas/EmbedField"
...
    FileContent:
      type: object
      required: [path, content, lines]
      properties:
        path:
          type: string
        content:
          type: string
        lines:
          type: integer
        language:
          type: string
    FileSearchResult:
      type: object
      required: [path, line, content]
      properties:
        path:
          type: string
        line:
          type: integer
        content:
          type: string
        matchStart:
          type: integer
        matchEnd:
          type: integer
...