# lib/api-zod/src/generated/types/embedField.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lowo Control Center API
 * OpenAPI spec version: 0.1.0
 */

export interface EmbedField {
  name: string;
  value: string;
  inline: boolean;
}